requirements.txt
streamlit==1.26.0
pandas==2.1.0
numpy==1.27.0
scikit-learn==1.3.0
scipy==1.11.0
Pillow==10.0.0
